Revert "fix(component): fix profile page component, docs, and tests"
