--- packages/registry/templates/pages/profile/__tests__/components/SocialLinksList.test.tsx
@@ -1,293 +1,0 @@
-// SocialLinksList.test.tsx
-import { render, screen, fireEvent } from "@testing-library/react";
-import { describe, it, expect, vi, beforeEach } from "vitest";
-import { SocialLinksList } from "../../components/SocialLinksList";
-
-// Mock dependencies
-vi.mock("../../../../utils/cn", () => ({
-    cn: (...classes: any[]) => classes.filter(Boolean).join(" "),
-}));
-
-vi.mock("@ignix-ui/button", () => ({
-    Button: ({ children, variant, size, onClick, className, animationVariant }: any) => (
-        <button
-            data-testid="button"
-            data-variant={variant}
-            data-size={size}
-            data-animation={animationVariant}
-            onClick={onClick}
-            className={className}
-        >
-            {children}
-        </button>
-    ),
-}));
-
-vi.mock("@ignix-ui/input", () => ({
-    AnimatedInput: ({ placeholder, variant, value, onChange }: any) => (
-        <input
-            data-testid="animated-input"
-            placeholder={placeholder}
-            data-variant={variant}
-            value={value}
-            onChange={(e) => onChange?.(e.target.value)}
-        />
-    ),
-}));
-
-vi.mock("@ignix-ui/typography", () => ({
-    Typography: ({ children, variant, weight, color, className }: any) => (
-        <div
-            data-variant={variant}
-            data-weight={weight}
-            data-color={color}
-            className={className}
-        >
-            {children}
-        </div>
-    ),
-}));
-
-vi.mock("lucide-react", () => ({
-    Link: () => <div data-testid="link-icon">LinkIcon</div>,
-    Twitter: () => <div data-testid="twitter-icon">TwitterIcon</div>,
-    Github: () => <div data-testid="github-icon">GithubIcon</div>,
-    Linkedin: () => <div data-testid="linkedin-icon">LinkedinIcon</div>,
-    Globe: () => <div data-testid="earth-icon">EarthIcon</div>,
-    Plus: () => <div data-testid="plus-icon">PlusIcon</div>,
-    Trash2: () => <div data-testid="trash-icon">TrashIcon</div>,
-}));
-
-// Mock the actual SocialLinksList component
-vi.mock("../../components/SocialLinksList", async () => {
-    const actual = await vi.importActual("../../components/SocialLinksList");
-    return {
-        ...actual,
-        SocialLinksList: (props: any) => {
-            const { links, isEditing, onLinksChange } = props;
-
-            if (isEditing) {
-                return (
-                    <div data-testid="social-links-editor">
-                        {links.map((link: any, index: number) => (
-                            <div key={link.id} data-testid={`social-link-${index}`}>
-                                <input
-                                    data-testid={`platform-input-${index}`}
-                                    value={link.platform}
-                                    onChange={(e) => {
-                                        const newLinks = [...links];
-                                        newLinks[index] = { ...newLinks[index], platform: e.target.value };
-                                        onLinksChange(newLinks);
-                                    }}
-                                    placeholder="Platform"
-                                />
-                                <input
-                                    data-testid={`url-input-${index}`}
-                                    value={link.url}
-                                    onChange={(e) => {
-                                        const newLinks = [...links];
-                                        newLinks[index] = { ...newLinks[index], url: e.target.value };
-                                        onLinksChange(newLinks);
-                                    }}
-                                    placeholder="https://..."
-                                />
-                                <button
-                                    data-testid={`remove-button-${index}`}
-                                    onClick={() => {
-                                        const newLinks = links.filter((_: any, i: number) => i !== index);
-                                        onLinksChange(newLinks);
-                                    }}
-                                >
-                                    <div data-testid="trash-icon">TrashIcon</div>
-                                </button>
-                            </div>
-                        ))}
-                        <button
-                            data-testid="add-button"
-                            onClick={() => {
-                                const newLinks = [...links, { id: Date.now().toString(), platform: '', url: '' }];
-                                onLinksChange(newLinks);
-                            }}
-                        >
-                            <div data-testid="plus-icon">PlusIcon</div>
-                            Add Social Link
-                        </button>
-                    </div>
-                );
-            }
-
-            // View mode
-            if (links.length === 0) {
-                return (
-                    <div data-testid="no-links-message">
-                        No social links added
-                    </div>
-                );
-            }
-
-            return (
-                <div data-testid="social-links-view">
-                    {links.map((link: any) => (
-                        <div key={link.id} data-testid={`social-link-${link.platform || 'unnamed'}`}>
-                            <div data-testid="link-icon">LinkIcon</div>
-                            <div>
-                                <div>{link.platform || 'Unnamed'}</div>
-                                <a
-                                    href={link.url}
-                                    target="_blank"
-                                    rel="noopener noreferrer"
-                                >
-                                    {link.url}
-                                </a>
-                            </div>
-                        </div>
-                    ))}
-                </div>
-            );
-        }
-    };
-});
-
-describe("SocialLinksList", () => {
-    const mockOnLinksChange = vi.fn();
-    const defaultLinks = [
-        { id: "1", platform: "Twitter", url: "https://twitter.com/johndoe" },
-        { id: "2", platform: "GitHub", url: "https://github.com/johndoe" },
-    ];
-
-    const defaultProps = {
-        links: defaultLinks,
-        isEditing: false,
-        onLinksChange: mockOnLinksChange,
-    };
-
-    beforeEach(() => {
-        mockOnLinksChange.mockClear();
-    });
-
-    it("renders social links in view mode", () => {
-        render(<SocialLinksList {...defaultProps} />);
-
-        expect(screen.getByTestId("social-links-view")).toBeInTheDocument();
-        expect(screen.getByText("Twitter")).toBeInTheDocument();
-        expect(screen.getByText("https://twitter.com/johndoe")).toBeInTheDocument();
-        expect(screen.getByText("GitHub")).toBeInTheDocument();
-        expect(screen.getByText("https://github.com/johndoe")).toBeInTheDocument();
-    });
-
-    it("shows 'No social links added' when links array is empty", () => {
-        render(<SocialLinksList {...defaultProps} links={[]} />);
-
-        expect(screen.getByTestId("no-links-message")).toBeInTheDocument();
-        expect(screen.getByText(/no social links added/i)).toBeInTheDocument();
-    });
-
-    it("renders links as clickable in view mode", () => {
-        render(<SocialLinksList {...defaultProps} />);
-
-        const links = screen.getAllByRole("link");
-        expect(links).toHaveLength(2);
-
-        expect(links[0]).toHaveAttribute("href", "https://twitter.com/johndoe");
-        expect(links[0]).toHaveAttribute("target", "_blank");
-        expect(links[0]).toHaveAttribute("rel", "noopener noreferrer");
-    });
-
-    it("shows link icon in view mode", () => {
-        render(<SocialLinksList {...defaultProps} />);
-
-        expect(screen.getAllByTestId("link-icon")).toHaveLength(2);
-    });
-
-    it("renders edit inputs when isEditing is true", () => {
-        render(<SocialLinksList {...defaultProps} isEditing={true} />);
-
-        expect(screen.getByTestId("social-links-editor")).toBeInTheDocument();
-
-        const platformInputs = screen.getAllByTestId(/^platform-input-/);
-        const urlInputs = screen.getAllByTestId(/^url-input-/);
-
-        expect(platformInputs).toHaveLength(2);
-        expect(urlInputs).toHaveLength(2);
-
-        expect(platformInputs[0]).toHaveValue("Twitter");
-        expect(urlInputs[0]).toHaveValue("https://twitter.com/johndoe");
-    });
-
-    it("calls onLinksChange when platform input changes", async () => {
-        render(<SocialLinksList {...defaultProps} isEditing={true} />);
-
-        const platformInput = screen.getByTestId("platform-input-0");
-
-        // Instead of using user.clear(), use fireEvent to set the value directly
-        fireEvent.change(platformInput, { target: { value: 'X' } });
-
-        // Check that onLinksChange was called with the correct value
-        expect(mockOnLinksChange).toHaveBeenCalledWith([
-            { id: "1", platform: "X", url: "https://twitter.com/johndoe" },
-            { id: "2", platform: "GitHub", url: "https://github.com/johndoe" },
-        ]);
-    });
-
-    it("calls onLinksChange when URL input changes", async () => {
-        render(<SocialLinksList {...defaultProps} isEditing={true} />);
-
-        const urlInput = screen.getByTestId("url-input-0");
-
-        // Instead of using user.clear() and user.type(), use fireEvent to set the value directly
-        fireEvent.change(urlInput, { target: { value: 'https://x.com/johndoe' } });
-
-        expect(mockOnLinksChange).toHaveBeenCalledWith([
-            { id: "1", platform: "Twitter", url: "https://x.com/johndoe" },
-            { id: "2", platform: "GitHub", url: "https://github.com/johndoe" },
-        ]);
-    });
-
-    it("adds new link when add button is clicked", () => {
-        render(<SocialLinksList {...defaultProps} isEditing={true} />);
-
-        const addButton = screen.getByTestId("add-button");
-        fireEvent.click(addButton);
-
-        expect(mockOnLinksChange).toHaveBeenCalledWith([
-            ...defaultLinks,
-            expect.objectContaining({
-                id: expect.any(String),
-                platform: '',
-                url: '',
-            }),
-        ]);
-    });
-
-    it("removes link when trash button is clicked", () => {
-        render(<SocialLinksList {...defaultProps} isEditing={true} />);
-
-        const removeButton = screen.getByTestId("remove-button-0");
-        fireEvent.click(removeButton);
-
-        expect(mockOnLinksChange).toHaveBeenCalledWith([defaultLinks[1]]);
-    });
-
-    it("shows 'Unnamed' for links without platform in view mode", () => {
-        const linksWithEmptyPlatform = [
-            { id: "1", platform: "", url: "https://example.com" },
-        ];
-
-        render(
-            <SocialLinksList
-                {...defaultProps}
-                links={linksWithEmptyPlatform}
-            />
-        );
-
-        expect(screen.getByText("Unnamed")).toBeInTheDocument();
-    });
-
-    it("shows trash and plus icons in edit mode", () => {
-        render(<SocialLinksList {...defaultProps} isEditing={true} />);
-
-        expect(screen.getByTestId("plus-icon")).toBeInTheDocument();
-        expect(screen.getAllByTestId("trash-icon")).toHaveLength(2);
-    });
-
-});
